Code refactored, added main menu, 2 new simulations, and a bunch of minor things
Code was refactored and divided into sensible parts, it's still a bit hacky and lacks documentation, will be updated in the future.
Added a main menu from which simulations can be accessed.
Added 2 new simulations in which the agent is trying to stay on course while estimating it's position using Kalman filter.
Added buttons for reseting and ending a simulation. On end, a plot of measurement and estimation errors will be drawn.
Added the credits screen.
Still need to incorporate the constant velocity simulation into the main program.